@@ -1,15 +1,15 @@
 name: notustohtml
 description: Convert the Notus document format to an HTML representation.
-version: 0.0.9-dev.3.0
+version: 1.0.0-rc.1
 homepage: https://github.com/JacobWrenn/notustohtml
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  notus: ^1.0.0-dev
-  quill_delta: ^2.0.0
-  html: ^0.14.0
+  notus: ^1.0.0-rc.1
+  quill_delta: ^3.0.0-nullsafety.1
+  html: ^0.15.0
 
 dev_dependencies:
   test:
