@@ -31,11 +31,11 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
     final iterator = DeltaIterator(input);
     final buffer = StringBuffer();
     final lineBuffer = StringBuffer();
-    NotusAttribute<String> currentBlockStyle;
+    NotusAttribute<String>? currentBlockStyle;
     var currentInlineStyle = NotusStyle();
     var currentBlockLines = [];
 
-    void _handleBlock(NotusAttribute<String> blockStyle) {
+    void _handleBlock(NotusAttribute<String>? blockStyle) {
       if (currentBlockLines.isEmpty) {
         return; // Empty block
       }
@@ -71,15 +71,15 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
       buffer.writeln();
     }
 
-    void _handleSpan(String text, Map<String, dynamic> attributes,
-        {bool hr = false, String source}) {
+    void _handleSpan(String text, Map<String, dynamic>? attributes,
+        {bool hr = false, String? source}) {
       final style = NotusStyle.fromJson(attributes);
       currentInlineStyle = _writeInline(
           lineBuffer, text, style, currentInlineStyle,
           hr: hr, source: source);
     }
 
-    void _handleLine(Map<String, dynamic> attributes) {
+    void _handleLine(Map<String, dynamic>? attributes) {
       final style = NotusStyle.fromJson(attributes);
       final lineBlock = style.get(NotusAttribute.block);
       if (lineBlock == currentBlockStyle) {
@@ -97,7 +97,7 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
     while (iterator.hasNext) {
       Operation op = iterator.next();
       bool hr = false;
-      String source;
+      String? source;
       if (op.data is BlockEmbed) {
         final embed = op.data as BlockEmbed;
         if (embed.type == "hr") {
@@ -120,7 +120,7 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
       final opText = op.data is String ? op.data as String : '';
       final lf = opText.indexOf('\n');
       if (lf == -1) {
-        _handleSpan(op.data, op.attributes, hr: hr, source: source);
+        _handleSpan(opText, op.attributes, hr: hr, source: source);
       } else {
         var span = StringBuffer();
         for (var i = 0; i < opText.length; i++) {
@@ -174,8 +174,8 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
 
   NotusStyle _writeInline(StringBuffer buffer, String text, NotusStyle style,
       NotusStyle currentStyle,
-      {bool hr, String source}) {
-    NotusAttribute wasA;
+      {bool? hr, String? source}) {
+    NotusAttribute? wasA;
     // First close any current styles if needed
     for (var value in currentStyle.values) {
       if (value.scope == NotusAttributeScope.line) continue;
@@ -204,7 +204,7 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
     if (source != null) {
       buffer.write("<img src=\"${source}\">");
     }
-    if (hr) {
+    if (hr!) {
       buffer.write("<hr>");
     }
     // Write the text itself
@@ -212,7 +212,7 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
     return style;
   }
 
-  void _writeAttribute(StringBuffer buffer, NotusAttribute attribute,
+  void _writeAttribute(StringBuffer buffer, NotusAttribute? attribute,
       {bool close = false}) {
     if (attribute == NotusAttribute.bold) {
       _writeBoldTag(buffer, close: close);
@@ -222,12 +222,14 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
       _writeTag(buffer, close: close, tag: "u");
     } else if (attribute == NotusAttribute.strikethrough) {
       _writeTag(buffer, close: close, tag: "del");
-    } else if (attribute.key == NotusAttribute.link.key) {
-      _writeLinkTag(buffer, attribute as NotusAttribute<String>, close: close);
+    } else if (attribute!.key == NotusAttribute.link.key) {
+      _writeLinkTag(buffer, attribute as NotusAttribute<String?>?,
+          close: close);
     } else if (attribute.key == NotusAttribute.heading.key) {
-      _writeHeadingTag(buffer, attribute as NotusAttribute<int>, close: close);
+      _writeHeadingTag(buffer, attribute as NotusAttribute<int?>, close: close);
     } else if (attribute.key == NotusAttribute.block.key) {
-      _writeBlockTag(buffer, attribute as NotusAttribute<String>, close: close);
+      _writeBlockTag(buffer, attribute as NotusAttribute<String?>?,
+          close: close);
     } else {
       throw ArgumentError('Cannot handle $attribute');
     }
@@ -237,7 +239,7 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
     buffer.write(!close ? "<$kBold>" : "</$kBold>");
   }
 
-  void _writeTag(StringBuffer buffer, {bool close = false, String tag}) {
+  void _writeTag(StringBuffer buffer, {bool close = false, String? tag}) {
     buffer.write(!close ? "<$tag>" : "</$tag>");
   }
 
@@ -245,22 +247,22 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
     buffer.write(!close ? "<$kItalic>" : "</$kItalic>");
   }
 
-  void _writeLinkTag(StringBuffer buffer, NotusAttribute<String> link,
+  void _writeLinkTag(StringBuffer buffer, NotusAttribute<String?>? link,
       {bool close = false}) {
     if (close) {
       buffer.write('</a>');
     } else {
-      buffer.write('<a href="${link.value}">');
+      buffer.write('<a href="${link!.value}">');
     }
   }
 
-  void _writeHeadingTag(StringBuffer buffer, NotusAttribute<int> heading,
+  void _writeHeadingTag(StringBuffer buffer, NotusAttribute<int?> heading,
       {bool close = false}) {
     var level = heading.value;
     buffer.write(!close ? "<h$level>" : "</h$level>");
   }
 
-  void _writeBlockTag(StringBuffer buffer, NotusAttribute<String> block,
+  void _writeBlockTag(StringBuffer buffer, NotusAttribute<String?>? block,
       {bool close = false}) {
     if (block == NotusAttribute.code) {
       if (!close) {
@@ -270,9 +272,9 @@ class _NotusHtmlEncoder extends Converter<Delta, String> {
       }
     } else {
       if (!close) {
-        buffer.write('<${kSimpleBlocks[block]}>');
+        buffer.write('<${kSimpleBlocks[block!]}>');
       } else {
-        buffer.write('</${kSimpleBlocks[block]}>');
+        buffer.write('</${kSimpleBlocks[block!]}>');
       }
     }
   }
@@ -284,9 +286,10 @@ class _NotusHtmlDecoder extends Converter<String, Delta> {
     Delta delta = Delta();
     Document html = parse(input);
 
-    html.body.nodes.asMap().forEach((index, node) {
+    html.body!.nodes.asMap().forEach((index, node) {
       var next;
-      if (index + 1 < html.body.nodes.length) next = html.body.nodes[index + 1];
+      if (index + 1 < html.body!.nodes.length)
+        next = html.body!.nodes[index + 1];
       delta = _parseNode(node, delta, next);
     });
     final text = delta.last.data is String ? delta.last.data as String : '';
@@ -302,30 +305,28 @@ class _NotusHtmlDecoder extends Converter<String, Delta> {
       if (element.localName == "ul") {
         element.children.forEach((child) {
           delta = _parseElement(
-              child, delta, _supportedElements[child.localName],
+              child, delta, _supportedElements[child.localName!],
               listType: "ul",
               next: next,
               isNewLine: isNewLine,
               inBlock: inBlock);
-          return delta;
         });
       }
       if (element.localName == "ol") {
         element.children.forEach((child) {
           delta = _parseElement(
-              child, delta, _supportedElements[child.localName],
+              child, delta, _supportedElements[child.localName!],
               listType: "ol",
               next: next,
               isNewLine: isNewLine,
               inBlock: inBlock);
-          return delta;
         });
       }
-      if (_supportedElements[element.localName] == null) {
+      if (_supportedElements[element.localName!] == null) {
         return delta;
       }
       delta = _parseElement(
-          element, delta, _supportedElements[element.localName],
+          element, delta, _supportedElements[element.localName!],
           next: next, isNewLine: isNewLine, inBlock: inBlock);
       return delta;
     } else {
@@ -341,9 +342,9 @@ class _NotusHtmlDecoder extends Converter<String, Delta> {
     }
   }
 
-  Delta _parseElement(Element element, Delta delta, String type,
-      {Map<String, dynamic> attributes,
-      String listType,
+  Delta _parseElement(Element element, Delta delta, String? type,
+      {Map<String, dynamic>? attributes,
+      String? listType,
       next,
       isNewLine,
       inBlock}) {
@@ -384,7 +385,7 @@ class _NotusHtmlDecoder extends Converter<String, Delta> {
     } else if (type == "embed") {
       if (element.localName == "img") {
         delta = delta
-          ..insert(BlockEmbed.image(element.attributes["src"]).toJson());
+          ..insert(BlockEmbed.image(element.attributes["src"]!).toJson());
       }
       if (element.localName == "hr") {
         delta = delta..insert(BlockEmbed.horizontalRule.toJson());
@@ -424,7 +425,7 @@ class _NotusHtmlDecoder extends Converter<String, Delta> {
       } else {
         element.nodes.forEach((node) {
           if (node.runtimeType == Element) {
-            var elementType = _supportedElements[(node as Element).localName];
+            var elementType = _supportedElements[(node as Element).localName!];
             if (elementType == null) {
               return;
             }
